requirements.txt
aiofiles==23.2.1
altair==5.2.0
amqp==5.2.0
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.2.0
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
channels==4.0.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
contourpy==1.2.0
cryptography==41.0.7
cycler==0.12.1
daphne==4.0.0
Django==4.2.5
django-autoslug==1.9.9
django-cors-headers==4.2.0
django-extensions==3.2.3
django-filter==23.3
django-restframework==0.0.1
django-solo==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
et-xmlfile==1.1.0
exceptiongroup==1.1.3
fastapi==0.109.0
ffmpy==0.3.1
filelock==3.13.1
fonttools==4.47.2
fsspec==2023.12.2
gradio==4.14.0
gradio_client==0.8.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
huggingface-hub==0.20.2
hyperlink==21.0.0
idna==3.6
importlib-resources==6.1.1
incremental==22.10.0
iniconfig==2.0.0
Jinja2==3.1.3
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
kombu==5.3.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
mdurl==0.1.2
msgpack==1.0.7
numpy==1.26.3
openpyxl==3.1.2
orjson==3.9.10
packaging==23.2
pandas==2.1.4
Pillow==10.0.1
pluggy==1.3.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.7
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pydub==0.25.1
Pygments==2.17.2
PyJWT==2.8.0
pyOpenSSL==23.3.0
pyparsing==3.1.1
pytest==7.4.2
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
referencing==0.32.1
requests==2.31.0
rich==13.7.0
rpds-py==0.17.1
semantic-version==2.10.0
service-identity==23.1.0
shellingham==1.5.4
singleton==0.1.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
starlette==0.35.1
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.0
tqdm==4.66.1
Twisted==23.10.0
txaio==23.1.1
typer==0.9.0
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
uvicorn==0.25.0
vine==5.1.0
wcwidth==0.2.13
websockets==11.0.3
whitenoise==6.6.0
zope.interface==6.1
